// src/sdk/generated/golf.d.ts
export interface components {
    schemas: {
        Unauthorized: {
            HttpStatusCode?: number;
            Code?: number;
            Description?: string;
            Help?: string;
        };
        Tournament: {
            /** @description The unique ID of the tournament */
            TournamentID?: number;
            /** @description The name of the tournament */
            Name?: string | null;
            /** @description The start date of this tournament */
            StartDate?: string | null;
            /** @description The end date of this tournament */
            EndDate?: string | null;
            /** @description Indicates whether this tournament is over */
            IsOver?: boolean;
            /** @description Indicates whether this tournament is in progress */
            IsInProgress?: boolean;
            /** @description The name of the venue where this tournament takes place */
            Venue?: string | null;
            /** @description The location where this tournament is played */
            Location?: string | null;
            /** @description The 18-hole par of this course */
            Par?: number | null;
            /** @description The total course yardage */
            Yards?: number | null;
            /** @description The total prize money of this tournament */
            Purse?: number | null;
            /** @description The first tee time of the upcoming round of this tournament (in US Eastern Time) */
            StartDateTime?: string | null;
            /** @description Indicates whether this tournament is canceled or not */
            Canceled?: boolean | null;
            /** @description Indicates whether this tournament is covered. Note: only stroke format tournaments are covered */
            Covered?: boolean | null;
            /** @description The city where this course is located */
            City?: string | null;
            /** @description The state where this course is located */
            State?: string | null;
            /** @description The zip code where this course is located */
            ZipCode?: string | null;
            /** @description The country where this course is located */
            Country?: string | null;
            /** @description The description of the time zone in which the course is located */
            TimeZone?: string | null;
            /** @description The format of the tournament (possible values: Stroke, Match, Team, Stableford) */
            Format?: string | null;
            /** @description The details of the rounds of this tournament */
            Rounds?: components["schemas"]["Round"][];
            /** @description The unique ID of this tournament in the SportRadar API */
            SportRadarTournamentID?: string | null;
            /** @description Shows the type of odds coverage for the tournament, with possible values being 'None'; 'Limited'; or 'Full' */
            OddsCoverage?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}